+++ MyMoney/MyMoney.Core/Data/Budget.cs
@@ -1,7 +1,6 @@
 ﻿using System;
 
-
-namespace MyMoney.Core.Model
+namespace MyMoney.Core.Data
 {
     public class Budget
     {
