+++ MyMoney/MyMoney.Core/Repositories/ITransactionRepository.cs
@@ -1,4 +1,4 @@
-﻿using MyMoney.Core.Model;
+﻿using MyMoney.Core.Data;
 using System;
 using System.Collections.Generic;
